Enabled Debian Bullseye in the build system

+++ build.sh
@@ -15,6 +15,13 @@ TARGET=""
 SAI_INTERFACE="redis"
 BASE_OS="buster"
 
+declare -A base_os_map
+base_os_map["deb10"]="buster"
+base_os_map["buster"]="buster"
+base_os_map["deb11"]="bullseye"
+base_os_map["bullseye"]="bullseye"
+
+
 print-help() {
     echo
     echo "$(basename ""$0"") [OPTIONS]"
@@ -72,6 +79,13 @@ if [[ "${IMAGE_TYPE}" != "standalone" && \
     exit 1
 fi
 
+if [ ! -v base_os_map["${BASE_OS}"] ]; then
+    echo "Unknown base OS \"${BASE_OS}\""
+    exit 1
+fi
+
+BASE_OS="${base_os_map[${BASE_OS}]}"
+
 # Clean the previous build
 rm -rf .build/
 
@@ -125,16 +139,16 @@ trap print-build-options EXIT
 
 # Build base Docker image
 if [ "${IMAGE_TYPE}" = "standalone" ]; then
-    docker build -f dockerfiles/${BASE_OS}/Dockerfile -t sc-base .
+    docker build -f dockerfiles/${BASE_OS}/Dockerfile -t sc-base:${BASE_OS} .
 elif [ "${IMAGE_TYPE}" = "server" ]; then
     find ${ASIC_PATH}/../ -type f -name \*.py -exec install -D {} .build/{} \;
     find ${ASIC_PATH}/../ -type f -name \*.json -exec install -D {} .build/{} \;
-    docker build -f dockerfiles/${BASE_OS}/Dockerfile.server -t sc-server-base .
+    docker build -f dockerfiles/${BASE_OS}/Dockerfile.server -t sc-server-base:${BASE_OS} .
     rm -rf .build/
 else
-    docker build -f dockerfiles/${BASE_OS}/Dockerfile.client -t sc-client .
+    docker build -f dockerfiles/${BASE_OS}/Dockerfile.client -t sc-client:${BASE_OS} .
     if [ "${SAI_INTERFACE}" = "thrift" ]; then
-        docker build -f dockerfiles/${BASE_OS}/Dockerfile.saithrift-client -t sc-thrift-client .
+        docker build -f dockerfiles/${BASE_OS}/Dockerfile.saithrift-client -t sc-thrift-client:${BASE_OS} .
     fi
 fi
 
@@ -143,11 +157,11 @@ pushd "${ASIC_PATH}/${TARGET}"
 IMG_NAME=$(echo "${ASIC_TYPE}-${TARGET}" | tr '[:upper:]' '[:lower:]')
 if [ "${IMAGE_TYPE}" = "standalone" ]; then
     if [ "${SAI_INTERFACE}" = "thrift" ]; then
-        docker build -f Dockerfile.saithrift -t sc-thrift-${IMG_NAME} .
+        docker build -f Dockerfile.saithrift --build-arg BASE_OS=${BASE_OS} -t sc-thrift-${IMG_NAME}:${BASE_OS} .
     else
-        docker build -f Dockerfile -t sc-${IMG_NAME} .
+        docker build -f Dockerfile --build-arg BASE_OS=${BASE_OS} -t sc-${IMG_NAME}:${BASE_OS} .
     fi
 elif [ "${IMAGE_TYPE}" = "server" ]; then
-    docker build -f Dockerfile.server -t sc-server-${IMG_NAME} .
+    docker build -f Dockerfile.server --build-arg BASE_OS=${BASE_OS} -t sc-server-${IMG_NAME}:${BASE_OS} .
 fi
 popd
